app/requirements.txt
django>=4.0.4
psycopg2-binary==2.9.1
cryptography==3.3.2
django-allauth>=0.43,<0.44
django-grappelli
pillow
django-ckeditor
django-extensions
django-crispy-forms
crispy-bootstrap5
django-cleanup
django-bootstrap-v5
django-htmx
django-jazzmin
django-filter
factory_boy
django-renderpdf
flake8
weasyprint
black